Use our own safe_makedirs() instead of os.makedirs().
os.makedirs() has a bug in that it does not understand paths that use both
'\' and '/' as path separators, because it decides once and for all that we
are on a POSIX system or an NT system and there’s no easy going back.

To avoid the problem we change all path separators to os.sep.