Fixes for upload-fileset
Uses a temporary directory when not uploading to an archive.
Ensures that there are no duplicate in files to upload.
Handles no configuration in arguments.
Fixes file mapper used for the torrent when not uploading an archive.
Changes error handling to use exceptions instead of returned booleans.
Moves all zipfile code to the archive method.
Refactors torrent bttree creation to be more readable.