Update check status command
Adds nimp environment to displayed information.
Adds free-parameters to check argument effects on the nimp environment.
Uses json to get a nice environment output without custom logic.
Rename python section to system and update it to use direct getters instead of the uname tuple.
Homogenize prints to use the percent format, in the same way as the logging.