Rework fileset stash commands
Creates a directory for each stash instead of storing all the files at the same place.
Includes more parameters in the stash name to reduce collisions.
Raises exception on error instead of returning false.
Ensures the parent directory exists when unstashing a file.
Uses a generic easy to find name for the stash summary file.